feat: checking for nonexistent bootstrap file

+++ persistence.sh
@@ -97,17 +97,20 @@ upsert_file() {
 }
 
 save_ansible_vars() {
-	[ -d ${ANSIBLE_DIR} ] && git clone ${BRAND_ANSIBLE_URL} ${ANSIBLE_DIR} 
+	[ -d ${ANSIBLE_DIR} ] || git clone ${BRAND_ANSIBLE_URL} ${ANSIBLE_DIR} 
 	now=$(date +"%m_%d_%y")
 	# Prepend bootstrap.log into ceremony.log before github commits and before it's copied to volumes/
-	COMBINED=$(cat ~/bootstrap.log; cat ${LOG_FILE})
-	echo "$COMBINED" > ${LOG_FILE}
-	echo "Finished: $(date)" >> "${now}_${LOG_FILE}"
+	if [ -f "${HOME}/bootstrap.log" ]; then
+		COMBINED=$(cat ~/bootstrap.log; cat ${LOG_FILE})
+		echo "$COMBINED" > ${LOG_FILE}
+	fi
+
+	echo "Finished: $(date)" >> "${LOG_FILE}"
 	## bootstrap.log will already have a Started: timestamp
-	cp ${LOG_FILE} "${ANSIBLE_DIR}/${now}_ceremony.log"
+	cp ${LOG_FILE} "${ANSIBLE_DIR}/ceremony_${now}.log"
 	git config --global user.name "ceremony-script"
 	git config --global user.email "ceremony@email.com"
-	git -C ${ANSIBLE_DIR}/ checkout -B "${now}-ceremony-artifacts"
+	git -C ${ANSIBLE_DIR}/ checkout -B "ceremony-artifacts-${now}"
 	git -C ${ANSIBLE_DIR}/ add ${ANSIBLE_DIR}/ &>> ${LOG_FILE}
 	git -C ${ANSIBLE_DIR}/ commit -m "Committing produced artifacts"
 	git -C ${ANSIBLE_DIR}/ push origin HEAD --force --porcelain &>> ${LOG_FILE}
